backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# HTTP client for Zoho API calls
httpx==0.26.0
aiohttp==3.9.3

# Environment and configuration
python-dotenv==1.0.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Database (optional - for token persistence)
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Security
python-jose[cryptography]==3.3.0

# AWS Bedrock for LLM
boto3==1.34.50
aioboto3==12.3.0

# Vector search and embeddings
numpy==1.26.4
faiss-cpu==1.7.4
pandas==2.2.0
openpyxl==3.1.2

# Web scraping
beautifulsoup4==4.12.3

# Document parsing
pypdf==4.0.1
python-docx==1.1.0
python-pptx==0.6.23

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0